+++ test/cases/hello-world-app/webpack.config.js
@@ -19,7 +19,6 @@ module.exports = {
   output: {
     path: path.join(__dirname, 'public'),
     publicPath: '/',
-    filename: 'assets/js/[name].[contenthash:8].js',
   },
 
   entry: {
@@ -28,7 +27,10 @@ module.exports = {
 
   plugins: [
     new PugPlugin({
-      extractCss: {
+      js: {
+        filename: 'assets/js/[name].[contenthash:8].js',
+      },
+      css: {
         filename: 'assets/css/[name].[contenthash:8].css',
       },
     }),