+++ test/cases/require-assets-method-render/webpack.config.js
@@ -15,7 +15,7 @@ module.exports = {
 
   plugins: [
     new PugPlugin({
-      extractCss: {
+      css: {
         filename: 'assets/css/[name].[contenthash:8].css',
       },
     }),