+++ test/cases/resolve-modules/webpack.config.js
@@ -21,7 +21,7 @@ module.exports = {
   plugins: [
     new PugPlugin({
       pretty: true,
-      extractCss: {
+      css: {
         filename: 'assets/css/[name].[contenthash:8].css',
       },
     }),